Use g_format_size()
	* libbalsa/libbalsa-vfs.c (libbalsa_vfs_get_size): get size as
	guint64.
	* libbalsa/libbalsa-vfs.h: ditto.
	* src/sendmsg-window.c (add_attachment), (add_urlref_attachment),
	(create_info_pane): ditto;
	(render_attach_size): use g_format_size() instead of deprecated
	g_format_size_for_display().

+++ libbalsa/libbalsa-vfs.h
@@ -71,7 +71,7 @@ const gchar * libbalsa_vfs_get_basename_utf8(const LibbalsaVfs * file);
 const gchar * libbalsa_vfs_get_mime_type(const LibbalsaVfs * file);
 const gchar * libbalsa_vfs_get_charset(const LibbalsaVfs * file);
 LibBalsaTextAttribute libbalsa_vfs_get_text_attr(const LibbalsaVfs * file);
-gsize libbalsa_vfs_get_size(const LibbalsaVfs * file);
+guint64 libbalsa_vfs_get_size(const LibbalsaVfs * file);
 GMimeStream * libbalsa_vfs_create_stream(const LibbalsaVfs * file,
                                          mode_t mode, 
                                          gboolean rdwr,
